backend/requirements.txt
# Requirements for the backend application
Django==5.1.3
djangorestframework==3.14.0
Pillow==10.4.0
drf-spectacular==0.28.0
asgiref==3.8.1
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-cors-headers==4.3.0
python-dotenv
django-filter